Milestone 1 Added: Blue - Sword skill (basic attack) Red - Fireball (projectile) Yellow - Lighting strike (ray from player to X tiles in front) Crates - Changing from square to circle sprite for now
Unit Stats
Player
Enemy

Testing skills
Skill 1
Skill 2
Skill 3

+++ Pleiades/Assets/Scripts/Unit.cs
@@ -6,8 +6,8 @@ public class Unit : MonoBehaviour
 {
     public float curHp;
     public float maxHp;
-    public float atkDmg;
-    public float damage;
+    public int atkDmg;
+    public int damage;
 
 
     public float Attack()
